Restore image in emails
I accidentally deleted the image on Cloudinary. While it's possible to recreate one with the same ID, it's impossible to recreate the same version number. So, this change removes the version number from the URL. This change means that if we modify the logo, it should be a new asset rather than a new version.