Prevent jobs from being started but not being processed
We've seen some jobs stall. It turns out that by default `job.moveToCompleted` fetches the next job. This change turns that off, so that jobs aren't marked as being started when we're not processing them.

It's turned off in `job.moveToFailed` by default, but this change also makes it explicit.

Refs 38450887d96cff4fe7b47230df3afe56589431d9, https://api.docs.bullmq.io/classes/v5.Job.html#moveToCompleted